Move `check_removed_reclass_variables()` to `inventory/lint.py`
