+++ tests/test_compile.py
@@ -1,6 +1,5 @@
 import click
 import pytest
-import yaml
 
 from pathlib import Path as P
 from typing import Dict
@@ -163,14 +162,3 @@ def test_check_parameters_component_versions(cluster_params: Dict, raises: bool)
 
     else:
         compile.check_parameters_component_versions(cluster_params)
-
-
-def test_check_removed_reclass_variables_error(tmp_path, config):
-    testf = tmp_path / "test.yml"
-    with open(testf, "w") as f:
-        yaml.safe_dump({"parameters": {"test": "${customer:name}"}}, f)
-
-    with pytest.raises(click.ClickException) as e:
-        compile._check_removed_reclass_variables(config, "tests", [testf])
-
-    assert "Found 1 usages of removed reclass variables in the tests." in str(e.value)
